Upsert Stored Procedure
with this SP, you can merge production table with using source table. 
The columns that is reasonable to have changes during the time is "LastLoginDate" and "PurchaseTotal". (I assumed that age is dateOfBirth , so we shouldn't update it every year.)
The rows which have changes on "LastLoginDate" or "PurchaseTotal" columns will be updated. and the userIDs that are not in prod.userID will be inserted.
there is line that is commented which created a Temp table to store the rows which is changed.